Update Xamarin project formats to use MSBuild SDK Extras
