+++ src/osuTK.Android/AndroidGraphicsMode.cs
