+++ src/osuTK.Android/Enums.cs
