+++ src/osuTK.Android/GameViewBase.cs
