+++ src/osuTK/Platform/Egl/EglWindowInfo.cs
@@ -30,7 +30,7 @@ using osuTK.Graphics;
 namespace osuTK.Platform.Egl
 {
     // Holds information about an EGL window.
-    internal class EglWindowInfo : IWindowInfo
+    public class EglWindowInfo : IWindowInfo
     {
         private IntPtr surface;
         private bool disposed;
