--- src/osuTK/Platform/iPhoneOS/iPhoneFactory.cs
@@ -1,87 +1,0 @@
-//
-// The Open Toolkit Library License
-//
-// Copyright (c) 2006 - 2009 the Open Toolkit library.
-// Copyright 2013 Xamarin Inc
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights to
-// use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
-// the Software, and to permit persons to whom the Software is furnished to do
-// so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-// OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
-// HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-// WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
-// OTHER DEALINGS IN THE SOFTWARE.
-//
-
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-using OpenGLES;
-
-using osuTK.Graphics;
-
-namespace osuTK.Platform.iPhoneOS
-{
-    internal class iPhoneFactory : PlatformFactoryBase
-    {
-        public override IGraphicsContext CreateGLContext(GraphicsMode mode, IWindowInfo window, IGraphicsContext shareContext, bool directRendering, int major, int minor, GraphicsContextFlags flags)
-        {
-            return new iPhoneOS.iPhoneOSGraphicsContext(mode, window, shareContext, major, minor, flags);
-        }
-
-        public override IGraphicsContext CreateGLContext(ContextHandle handle, IWindowInfo window, IGraphicsContext shareContext, bool directRendering, int major, int minor, GraphicsContextFlags flags)
-        {
-            return new iPhoneOS.iPhoneOSGraphicsContext(handle, window, shareContext, major, minor, flags);
-        }
-
-        public override GraphicsContext.GetCurrentContextDelegate CreateGetCurrentGraphicsContext()
-        {
-            return () => {
-                EAGLContext c = EAGLContext.CurrentContext;
-                IntPtr h = IntPtr.Zero;
-                if (c != null)
-                {
-                    h = c.Handle;
-                }
-                return new ContextHandle(h);
-            };
-        }
-
-        public override INativeWindow CreateNativeWindow(int x, int y, int width, int height, string title, GraphicsMode mode, GameWindowFlags options, DisplayDevice device)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override IDisplayDeviceDriver CreateDisplayDeviceDriver()
-        {
-            return new AndroidDisplayDeviceDriver();
-        }
-
-        public override osuTK.Input.IKeyboardDriver2 CreateKeyboardDriver()
-        {
-            throw new NotImplementedException();
-        }
-
-        public override osuTK.Input.IMouseDriver2 CreateMouseDriver()
-        {
-            throw new NotImplementedException();
-        }
-
-        public override osuTK.Input.IJoystickDriver2 CreateJoystickDriver()
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
